Tracked down and fixed up memory corruption issues (due to a wrongly calculated Alignment::num_states) when processing a DNA input with a YAML Protein model, via "-st NT2AA" on the command-line.  The bug fixed in 5(c), and the assumption removed in 7, were causing trouble. Though, it would be better if the functions changed in 7 worked in terms of IntVector rather than temporary arrays.
1. Changed the types of some icky members of Alignment:
   (a) non_stop_codon (is now an IntVector)
   (b) codon_table (is now an IntVector)
   (c) genetic_code (is now a std::string)
   that makes copying of that data between Alignment
   instances a mere matter of... assignments;
2. Added Alignment::copyStateInfoFrom, which copies
   num_states, sequence_type, seq_type, STATE_UNKNOWN,
   genetic_code, codon_table, and non_stop_codon.
   This is now called from Alignment member functions:
   (a) Alignment::extractSubAlignment
   (b) Alignment::extractPatterns
   (c) Alignment::extractPatternFreqs
   (d) Alignment::createBootstrapAlignment
   (e) Alignment::createGapMaskedAlignment
   (f) Alignment::copyAlignment
3. SuperAlignment::concatenateAlignments can't
   use change 2 and does the copies directly.
4. Alignment constructors and destructors no longer
   need to initialize 1(a), 1(b), 1(c) to nullptr,
   or delete them.
5. Rewrote Alignment member functions (to use the
   members that now have different types):
   (a) Alignment::isStopCodon
   (b) Alignment::getNumNonstopCodon
   (c) Alignment::initCodon
       (This has been redone, so that it sets
        num_states to the number of proteins, if
        seq_type is to be SeqType::SEQ_PROTEIN, if it
        is passed an nt2aa parameter of ... true)
       (Formerly it was setting it to the number of
        non-stop codons, ee-uw) (this caused all sorts
        of problems in code that assumed STATE_UNKNOWN
        which was 23 was greater than num_states, which
        was 61 (it should have been 20)).
6. computeAdapter(), in main/phylotesting.cpp, no longer
   needs to cast the chars in an Alignment's codon_table
   (Because it is now an IntVector, no need to cast to int).
   Neither does Alignment::convertCodonToAA().
7. Alignment::countStatesForSites() - both versions -
   now count "out of range" states as unknown, rather than
   doing out-of-range writes (!).

+++ main/phylotesting.cpp
@@ -500,13 +500,13 @@ double computeAdapter(Alignment *orig_aln, Alignment *newaln, int &adjusted_df) 
     for (int aa = 0; aa < newaln->num_states; ++aa) {
         double sum = 0;
         for (int codon = 0; codon < orig_aln->num_states; codon++) {
-            if (newaln->convertState(orig_aln->genetic_code[(int)orig_aln->codon_table[codon]]) == aa) {
+            if (newaln->convertState(orig_aln->genetic_code[orig_aln->codon_table[codon]]) == aa) {
                 sum += codon_freq[codon];
             }
         }
         sum = 1.0/sum;
         for (int codon = 0; codon < orig_aln->num_states; codon++) {
-            if (newaln->convertState(orig_aln->genetic_code[(int)orig_aln->codon_table[codon]]) == aa) {
+            if (newaln->convertState(orig_aln->genetic_code[orig_aln->codon_table[codon]]) == aa) {
                 codon_freq[codon] *= sum;
             }
         }
@@ -521,7 +521,7 @@ double computeAdapter(Alignment *orig_aln, Alignment *newaln, int &adjusted_df) 
         if (codon_counts[codon] == 0) {
             continue;
         }
-        has_AA[newaln->convertState(orig_aln->genetic_code[(int)orig_aln->codon_table[codon]])] = true;
+        has_AA[newaln->convertState(orig_aln->genetic_code[orig_aln->codon_table[codon]])] = true;
         adapter += codon_counts[codon]*log(codon_freq[codon]);
         adjusted_df++;
     }
